t2i: Execute pipeline in subprocess
Currently, we run the pipeline in a thread to not block the main loop.
However, we can't kill a thread in Python to terminate the execution
of the pipeline.

Fix this by using a subprocess that is given the pipeline function to
execute. We use pipes to communicate.
Additionally, we implement a cancel button that allows us to
terminate the execution.